Edit: Moved context provider declaration to domain layer
