+++ src/shared/container/contexts/ThemeContext/contexts/useThemeContext.ts
@@ -1,5 +1,5 @@
 import { createContext, useContext } from 'react';
-import IThemeContext from 'shared/container/contexts/ThemeContext/models/IThemeContext';
+import IThemeContext from 'shared/domain/providers/IThemeContext';
 
 export const ThemeContextProvider = createContext<IThemeContext>(
   {} as IThemeContext,
